fix(config): write the config atomically from every writer, not just the API
Seven writers still replaced scoring_config.json with `open(path, 'w')` followed
by json.dump — truncate, then write. That was survivable while the file was
tracked: a torn write left something `git checkout` could restore. Untracking it
removed that net, so a crash mid-dump now destroys the operator's only copy of
their overrides, and it is the file that holds viewer.password,
users.*.password_hash, upload.password, frame.tokens and immich.api_key.
calibrate's three appliers and percentile_normalizer's took no lock either, so a
CLI run could also land inside a viewer write.

atomic_write_json already existed for exactly this — mkstemp beside the target,
fsync, chmod to the destination's preserved mode, os.replace, fsync the
directory — and the API and optimizer writers already used it. It could not be
shared: it lived in api.config, and config, calibrate and database reaching into
the api package would invert the dependency direction that default_config_path's
own docstring exists to preserve. It and its four helpers are stdlib-only, so
they move to config_resolve, which every one of those modules already imports;
api.config re-exports them, and the eight-line write_user_config wrapper moves
with them. All seven writers now go through it.

Two tests were patching that primitive at api.config, where it is now only
re-exported rather than looked up, which made both mocks inert — the exact
failure CLAUDE.md warns about for FastAPI dependencies, in a different guise.
The unwritable-config test would have reported a stale key as evicted; the lock
test would have recorded nothing and still passed. Both now patch the module the
name resolves in, and both were confirmed to go red under an injected fault.

The lock test also asserted that exactly one write happened, which was only ever
true because the other one was invisible. Building a ScoringConfig over a fixture
whose weights do not sum to 100 auto-corrects and saves, so the optimizer's apply
has always rewritten the file twice; the first write simply bypassed the writer
being watched. Both are under CONFIG_WRITE_LOCK, and the assertion now says so —
every write locked, at least one write seen — which pins the more interesting of
the two: a rewrite that happens during the read half of a read-modify-write.
